+++ ua-free-consent-manager/ua-free-consent-manager.php
@@ -5,7 +5,7 @@
  * License URI: https://www.gnu.org/licenses/gpl-2.0.html
  * Author URI: https://uafree.org/
  * Plugin URI: https://uafree.org/ua-free-plugins/#plugin-ua-free-consent-manager
- * Version: 0.1.6
+ * Version: 0.1.7
  * Author: UA FREE
  * Text Domain: ua-free-consent-manager
  * Domain Path: /languages
@@ -18,7 +18,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'UAFREE_CONSENT_MANAGER_VERSION', '0.1.6' );
+define( 'UAFREE_CONSENT_MANAGER_VERSION', '0.1.7' );
 define( 'UAFREE_CONSENT_MANAGER_FILE', __FILE__ );
 define( 'UAFREE_CONSENT_MANAGER_DIR', plugin_dir_path( __FILE__ ) );
 define( 'UAFREE_CONSENT_MANAGER_URL', plugin_dir_url( __FILE__ ) );